Fix error where `pg` is not defined on `pg.release()`
This means an error in pg.connect(), but the error gets swallowed
because the `finally` block tries to call `release()`

```
[_login:edf631c4] !! 500 TypeError: Cannot read property 'release' of
undefined
    at Object.default_1 [as default]
(/src/src/models/user/getByExternalAuth.js:20:7)
    at process._tickDomainCallback (internal/process/next_tick.js:135:7)
```
