+++ src/models/project/addUser.js
@@ -7,26 +7,20 @@ const pgPool = getPgPool();
 
 // opts: userId, projectId
 export default async function addUserToProject(opts) {
-  let pg;
-  try {
-    pg = await pgPool.connect();
-    const q = `insert into projectuser (
+  const q = `insert into projectuser (
       id, project_id, user_id
     ) values (
       $1, $2, $3
     )`;
-    const v = [
-      uuid.v4().replace(/-/g, ""),
-      opts.projectId,
-      opts.userId,
-    ];
-    pg.query(q, v);
+  const v = [
+    uuid.v4().replace(/-/g, ""),
+    opts.projectId,
+    opts.userId,
+  ];
+  await pgPool.query(q, v);
 
-    await populateEnvUser({
-      project_id: opts.projectId,
-      user_id: opts.userId,
-    });
-  } finally {
-    pg.release();
-  }
+  await populateEnvUser({
+    project_id: opts.projectId,
+    user_id: opts.userId,
+  });
 }
